align data types for SGFCNodeTraits and SGFCPropertyTraits with libsgfc++ (#14)
until now SgfcKit used "unsigned long long", the same as libsgfc++.
after the aligning with libsgfc++, SgfcKit now uses a uint64_t, again
the same as libsgfc++.

+++ include/SGFCPropertyTraits.h
@@ -1,5 +1,5 @@
 // -----------------------------------------------------------------------------
-// Copyright 2020 Patrick Näf (herzbube@herzbube.ch)
+// Copyright 2020-2026 Patrick Näf (herzbube@herzbube.ch)
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -49,7 +49,7 @@
 /// @note Using an unsigned numerical type as the underlying type of this
 /// typedef is important to make the bitwise operators, specifically the
 /// shift left and shift right operators, unambiguous and safe to use.
-typedef unsigned long long SGFCPropertyTraits;
+typedef uint64_t SGFCPropertyTraits;
 
 /// @brief SGFCPropertyTrait enumerates the individual traits that an
 /// SGFCProperty object can have. An SGFCProperty object can have several
